+++ src/index.ts
@@ -1,13 +1,13 @@
 import getRegistryName = require('encode-registry') // TODO: remove this. BREAKING CHANGE
+import {
+  DirectoryResolution,
+  PackageMeta,
+  Resolution,
+} from './createResolver'
 import createFetcher, {FetchedPackage, PackageContentInfo} from './fetch'
 import pkgIdToFilename from './fs/pkgIdToFilename'
 import {read, save, Store} from './fs/storeController'
 import pkgIsUntouched from './pkgIsUntouched'
-import resolve, {
-  DirectoryResolution,
-  PackageMeta,
-  Resolution,
-} from './resolve'
 import resolveStore from './resolveStore'
 
 export {
@@ -15,7 +15,6 @@ export {
   createFetcher,
   PackageContentInfo,
   FetchedPackage,
-  resolve,
   DirectoryResolution,
   Resolution,
   PackageMeta,
